+++ src/main/java/net/luminis/tls/engine/HostnameVerifier.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2020, 2021, 2022, 2023, 2024 Peter Doornbosch
+ * Copyright © 2019, 2020, 2021, 2022, 2023, 2024 Peter Doornbosch
  *
  * This file is part of Agent15, an implementation of TLS 1.3 in Java.
  *
@@ -16,18 +16,12 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package net.luminis.tls.handshake;
+package net.luminis.tls.engine;
 
-import java.io.IOException;
+import java.security.cert.X509Certificate;
 
+public interface HostnameVerifier {
 
-public interface ClientMessageSender {
+    boolean verify(String hostname, X509Certificate serverCertificate);
 
-    void send(ClientHello clientHello) throws IOException;
-
-    void send(FinishedMessage finishedMessage) throws IOException;
-
-    void send(CertificateMessage certificateMessage) throws IOException;
-
-    void send(CertificateVerifyMessage certificateVerifyMessage);
 }