+++ src/main/java/net/luminis/tls/engine/ServerMessageProcessor.java
@@ -16,10 +16,12 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package net.luminis.tls.handshake;
+package net.luminis.tls.engine;
 
 import net.luminis.tls.ProtectionKeysType;
 import net.luminis.tls.TlsProtocolException;
+import net.luminis.tls.engine.MessageProcessor;
+import net.luminis.tls.handshake.*;
 
 import java.io.IOException;