+++ src/main/java/net/luminis/tls/engine/TlsStatusEventHandler.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package net.luminis.tls.handshake;
+package net.luminis.tls.engine;
 
 import net.luminis.tls.NewSessionTicket;
 import net.luminis.tls.TlsProtocolException;