+++ src/main/java/net/luminis/tls/engine/TrafficSecrets.java
@@ -16,7 +16,18 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package net.luminis.tls;
+package net.luminis.tls.engine;
 
-public class Message {
+public interface TrafficSecrets {
+
+    byte[] getClientEarlyTrafficSecret();
+
+    byte[] getClientHandshakeTrafficSecret();
+
+    byte[] getServerHandshakeTrafficSecret();
+
+    byte[] getClientApplicationTrafficSecret();
+
+    byte[] getServerApplicationTrafficSecret();
+
 }