+++ src/test/java/net/luminis/tls/TlsStateTest.java
@@ -19,6 +19,8 @@
 package net.luminis.tls;
 
 import at.favre.lib.hkdf.HKDF;
+import net.luminis.tls.engine.impl.TlsState;
+import net.luminis.tls.engine.impl.TranscriptHash;
 import net.luminis.tls.util.ByteUtils;
 import net.luminis.tls.util.FieldGetter;
 import org.junit.jupiter.api.Test;